+++ sapphire/common/database/settings.py
@@ -1,0 +1,5 @@
+from pydantic import AnyUrl, BaseModel
+
+
+class BaseDatabaseSettings(BaseModel):
+    dsn: AnyUrl
