+++ sapphire/database/cli.py
@@ -1,103 +1,30 @@
-import asyncio
-from typing import Optional
-
 import typer
-from loguru import logger
-from rich.console import Console
-from rich.table import Table
 
-from sapphire.common.database.service import FixtureFormatEnum
+from sapphire.common.database.cli import get_fixtures_cli, get_migrations_cli
+from sapphire.common.utils.settings import get_settings
 
-from .service import BaseDatabaseService
-
-
-@logger.catch
-def migrations_list(ctx: typer.Context):
-    database_service: BaseDatabaseService = ctx.obj["database"]
-
-    database_service.show_migrations()
+from .service import get_service
+from .settings import Settings
 
 
-@logger.catch
-def migrations_apply(ctx: typer.Context):
-    database_service: BaseDatabaseService = ctx.obj["database"]
+def callback(ctx: typer.Context):
+    ctx.obj = ctx.obj or {}
 
-    database_service.migrate()
+    if settings := ctx.obj.get("settings"):
+        settings = settings.database
+    else:
+        settings = get_settings(Settings)
+    database_service = get_service(settings=settings)
 
-
-@logger.catch
-def migrations_rollback(
-    ctx: typer.Context,
-    revision: Optional[str] = typer.Argument(
-        None,
-        help="Revision id or relative revision (`-1`, `-2`)",
-    ),
-):
-    database_service: BaseDatabaseService = ctx.obj["database"]
-
-    database_service.rollback(revision=revision)
+    ctx.obj["settings"] = settings
+    ctx.obj["database"] = database_service
 
 
-@logger.catch
-def migrations_create(
-        ctx: typer.Context,
-        message: Optional[str] = typer.Option(
-            None,
-            "-m", "--message",
-            help="Migration short message",
-        ),
-):
-    database_service: BaseDatabaseService = ctx.obj["database"]
-
-    database_service.create_migration(message=message)
-
-
-def get_migrations_cli() -> typer.Typer:
-    cli = typer.Typer(name="Migration")
-
-    cli.command(name="apply")(migrations_apply)
-    cli.command(name="rollback")(migrations_rollback)
-    cli.command(name="create")(migrations_create)
-    cli.command(name="list")(migrations_list)
-
-    return cli
-
-
-@logger.catch
-def fixtures_list(ctx: typer.Context):
-    database_service = ctx.obj["database"]
-
-    fixtures = database_service.get_fixtures()
-
-    table = Table("Name", "Format")
-    for fixture in fixtures:
-        table.add_row(fixture.name, fixture.format.value)
-    Console().print(table)
-
-
-@logger.catch
-def fixtures_apply(
-    ctx: typer.Context,
-    names: list[str] = typer.Argument(..., help="Fixtures names"),
-    fixture_format: FixtureFormatEnum = typer.Option(
-        FixtureFormatEnum.YAML,
-        "-f", "--format",
-        help="Fixture format/extension",
-    ),
-):
-    database_service = ctx.obj["database"]
-
-    loop = asyncio.get_event_loop()
-    for name in names:
-        loop.run_until_complete(database_service.apply_fixture(
-            name=name,
-            fixture_format=fixture_format,
-        ))
-
-
-def get_fixtures_cli() -> typer.Typer:
+def get_cli() -> typer.Typer:
     cli = typer.Typer()
-    cli.command(name="list")(fixtures_list)
-    cli.command(name="apply")(fixtures_apply)
+
+    cli.callback()(callback)
+    cli.add_typer(get_fixtures_cli(), name="fixtures")
+    cli.add_typer(get_migrations_cli(), name="migrations")
 
     return cli
