+++ sapphire/database/models/__init__.py
@@ -1,6 +1,69 @@
 from .base import Base
 from .messenger import Chat, ChatMember, Message
 from .notifications import Notification
-from .projects import Participant, ParticipantStatusEnum, Project, ProjectHistory, ProjectStatusEnum 
-from .storage import *
-from .users import *
+from .projects import (
+    Participant,
+    ParticipantStatusEnum,
+    Position,
+    PositionSkill,
+    Project,
+    ProjectHistory,
+    ProjectStatusEnum,
+    Review,
+)
+from .storage import Skill, Specialization, SpecializationGroup
+from .users import Profile, User, UserSkill
+
+
+MODELS = (
+    # Messenger
+    Chat,
+    ChatMember,
+    Message,
+    # Notifications
+    Notification,
+    # Projects
+    Participant,
+    Position,
+    PositionSkill,
+    Project,
+    ProjectHistory,
+    Review,
+    # Storage
+    Skill,
+    Specialization,
+    SpecializationGroup,
+    # Users
+    Profile,
+    User,
+    UserSkill,
+)
+__all__ = (
+    # Base
+    "Base",
+    # Messenger
+    "Chat",
+    "ChatMember",
+    "Message",
+    # Notifications
+    "Notification",
+    # Projects
+    "Participant",
+    "ParticipantStatusEnum",
+    "Position",
+    "PositionSkill",
+    "Project",
+    "ProjectHistory",
+    "ProjectStatusEnum",
+    "Review",
+    # Storage
+    "Skill",
+    "Specialization",
+    "SpecializationGroup",
+    # Users
+    "Profile",
+    "User",
+    "UserSkill",
+    # All models
+    "MODELS",
+)
