+++ sapphire/messenger/api/rest/chats/dependencies.py
@@ -5,8 +5,8 @@ import fastapi
 from sapphire.common.api.exceptions import HTTPForbidden, HTTPNotFound
 from sapphire.common.jwt.dependencies.rest import is_auth
 from sapphire.common.jwt.models import JWTData
+from sapphire.database.models import Chat
 from sapphire.messenger import database
-from sapphire.messenger.database.models import Chat
 
 
 async def get_path_chat(
