+++ sapphire/notifications/api/rest/notifications/dependencies.py
@@ -5,8 +5,8 @@ import fastapi
 from sapphire.common.api.exceptions import HTTPForbidden, HTTPNotFound
 from sapphire.common.jwt.dependencies.rest import is_auth
 from sapphire.common.jwt.models import JWTData
+from sapphire.database.models import Notification
 from sapphire.notifications import database
-from sapphire.notifications.database.models import Notification
 
 
 async def get_path_notification(
