+++ sapphire/notifications/api/rest/notifications/handlers.py
@@ -3,8 +3,8 @@ import fastapi
 from sapphire.common.api.dependencies.pagination import Pagination, pagination
 from sapphire.common.jwt.dependencies.rest import is_auth
 from sapphire.common.jwt.models import JWTData
+from sapphire.database.models import Notification
 from sapphire.notifications import database
-from sapphire.notifications.database.models import Notification
 
 from .dependencies import get_path_notification
 from .schemas import (
