+++ sapphire/projects/api/rest/positions/dependencies.py
@@ -5,8 +5,8 @@ import fastapi
 from sapphire.common.api.exceptions import HTTPForbidden, HTTPNotFound
 from sapphire.common.jwt.dependencies.rest import is_auth
 from sapphire.common.jwt.models import JWTData
+from sapphire.database.models import Position
 from sapphire.projects import database
-from sapphire.projects.database.models import Position
 
 
 async def get_path_position(
