+++ sapphire/projects/api/rest/positions/handlers.py
@@ -4,8 +4,8 @@ from sapphire.common.api.dependencies.pagination import Pagination, pagination
 from sapphire.common.api.exceptions import HTTPForbidden, HTTPNotFound
 from sapphire.common.jwt.dependencies.rest import is_auth
 from sapphire.common.jwt.models import JWTData
+from sapphire.database.models import Position
 from sapphire.projects import database
-from sapphire.projects.database.models import Position
 
 from .dependencies import get_path_position, path_position_is_owner
 from .schemas import (
