+++ sapphire/projects/api/rest/positions/skills/handlers.py
@@ -2,12 +2,12 @@ import uuid
 
 import fastapi
 
+from sapphire.database.models import Position
 from sapphire.projects import database
 from sapphire.projects.api.rest.positions.dependencies import (
     get_path_position,
     path_position_is_owner,
 )
-from sapphire.projects.database.models import Position
 
 
 async def update_position_skills(
