+++ sapphire/projects/api/rest/projects/handlers.py
@@ -8,9 +8,9 @@ from sapphire.common.api.dependencies.pagination import Pagination, pagination
 from sapphire.common.api.exceptions import HTTPForbidden
 from sapphire.common.jwt.dependencies.rest import is_activated
 from sapphire.common.jwt.models import JWTData
+from sapphire.database.models import Project
 from sapphire.projects import database
 from sapphire.projects.api.rest.schemas import ProjectResponse
-from sapphire.projects.database.models import Project
 
 from .dependencies import get_path_project, path_project_is_owner
 from .schemas import (
