+++ sapphire/settings.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
+from .database import Settings as DatabaseSettings
 from .email import Settings as EmailSettings
 from .messenger import Settings as MessengerSettings
 from .notifications import Settings as NotificationsSettings
@@ -9,6 +10,7 @@ from .users import Settings as UsersSettings
 
 
 class Settings(BaseModel):
+    database: DatabaseSettings = DatabaseSettings()
     email: EmailSettings = EmailSettings()
     messenger: MessengerSettings = MessengerSettings()
     notifications: NotificationsSettings = NotificationsSettings()
